5/2 14:32 배포 환경에서 recorder.js가 제대로 작동하지 않던 현상을 수정
