+++ package.json
@@ -12,6 +12,7 @@
     "@popperjs/core": "^2.11.7",
     "axios": "^1.3.5",
     "core-js": "^3.8.3",
+    "inline-worker": "^1.1.0",
     "register-service-worker": "^1.7.2",
     "roboto-fontface": "*",
     "v-calendar": "^3.0.3",
