+++ src/util/recorder/index.js
@@ -1,0 +1,1 @@
+module.exports = require("./recorder").Recorder;
